pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nova-scanner"
version = "3.1.0"
description = "NOVA v3.1 - The Forensic Code Auditor"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "NOVA Team"}
]
keywords = [
    "code-analysis",
    "forensics",
    "security",
    "static-analysis",
    "code-quality",
    "git",
    "cli"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: Security",
]

dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "GitPython>=3.1.0",
    "radon>=6.0.0",
    "bandit>=1.7.0",
    "pygount>=1.6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
full = [
    "langgraph>=0.0.50",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-community>=0.0.20",
    "qdrant-client>=1.7.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
nova = "nova.cli:main"

[project.urls]
Homepage = "https://github.com/Maca2024/NOVA-GIT-SCANNER_22"
Repository = "https://github.com/Maca2024/NOVA-GIT-SCANNER_22"
Issues = "https://github.com/Maca2024/NOVA-GIT-SCANNER_22/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["nova*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=nova --cov-report=term-missing"
